// projects/rpm-app/src/app/core/api/model/match.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Padel Ranking API
 *
 * API REST para gestión de rankings de pádel.
 *
 * ## Autenticación
 * Usar JWT Bearer token. Obtener el token mediante POST /api/auth/login.
 *
 * OpenAPI spec version: 1.0.0
 */
import type { MatchStatus } from './matchStatus';

export interface Match {
  id: string;
  rankingId: string;
  /** @nullable */
  rankingName?: string | null;
  /** @nullable */
  rankingStartDate?: string | null;
  /** @nullable */
  rankingEndDate?: string | null;
  teamOneId: string;
  /** @nullable */
  teamOneName?: string | null;
  /** @nullable */
  t1p1Name?: string | null;
  /** @nullable */
  t1p1Surname?: string | null;
  /** @nullable */
  t1p2Name?: string | null;
  /** @nullable */
  t1p2Surname?: string | null;
  teamTwoId: string;
  /** @nullable */
  teamTwoName?: string | null;
  /** @nullable */
  t2p1Name?: string | null;
  /** @nullable */
  t2p1Surname?: string | null;
  /** @nullable */
  t2p2Name?: string | null;
  /** @nullable */
  t2p2Surname?: string | null;
  /** @nullable */
  matchDate?: string | null;
  /** @nullable */
  resultSubmittedAt?: string | null;
  /** @nullable */
  resultSubmittedBy?: string | null;
  /** @nullable */
  submittedByEmail?: string | null;
  status: MatchStatus;
  /** @nullable */
  set1TeamOne?: number | null;
  /** @nullable */
  set1TeamTwo?: number | null;
  /** @nullable */
  set2TeamOne?: number | null;
  /** @nullable */
  set2TeamTwo?: number | null;
  /** @nullable */
  set3TeamOne?: number | null;
  /** @nullable */
  set3TeamTwo?: number | null;
  /** @nullable */
  gamesTeamOne?: number | null;
  /** @nullable */
  gamesTeamTwo?: number | null;
  /** @nullable */
  setsTeamOne?: number | null;
  /** @nullable */
  setsTeamTwo?: number | null;
  /** @nullable */
  pointsTeamOne?: number | null;
  /** @nullable */
  pointsTeamTwo?: number | null;
  /** @nullable */
  confirmedAt?: string | null;
  /** @nullable */
  confirmedBy?: string | null;
  /** @nullable */
  confirmedByEmail?: string | null;
  /** @nullable */
  disputedAt?: string | null;
  /** @nullable */
  disputedBy?: string | null;
  /** @nullable */
  disputedByEmail?: string | null;
  createdAt?: string;
  updatedAt?: string;
}