// projects/rpm-app/src/app/core/api/model/liveMatchSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Padel Ranking API
 *
 * API REST para gestión de rankings de pádel.
 *
 * ## Autenticación
 * Usar JWT Bearer token. Obtener el token mediante POST /api/auth/login.
 *
 * OpenAPI spec version: 1.0.0
 */
import type { LiveMatchSessionGameMode } from './liveMatchSessionGameMode';
import type { LiveMatchSessionIsTiebreak } from './liveMatchSessionIsTiebreak';
import type { LiveMatchSessionServingTeam } from './liveMatchSessionServingTeam';
import type { LiveMatchStatus } from './liveMatchStatus';

export interface LiveMatchSession {
  id: string;
  matchId: string;
  rankingId: string;
  /** @nullable */
  rankingName?: string | null;
  teamOneId: string;
  /** @nullable */
  teamOneName?: string | null;
  teamTwoId: string;
  /** @nullable */
  teamTwoName?: string | null;
  requestedBy?: string;
  /** @nullable */
  requestedByName?: string | null;
  /** @nullable */
  requestedByEmail?: string | null;
  /** @nullable */
  requestedByTeamId?: string | null;
  /** @nullable */
  requestedByTeamName?: string | null;
  status: LiveMatchStatus;
  currentSet: number;
  pointsTeamOne: string;
  pointsTeamTwo: string;
  set1TeamOne?: number;
  set1TeamTwo?: number;
  set2TeamOne?: number;
  set2TeamTwo?: number;
  /** @nullable */
  set3TeamOne?: number | null;
  /** @nullable */
  set3TeamTwo?: number | null;
  setsTeamOne?: number;
  setsTeamTwo?: number;
  gamesTeamOne?: number;
  gamesTeamTwo?: number;
  /** @nullable */
  gameMode?: LiveMatchSessionGameMode;
  servingTeam?: LiveMatchSessionServingTeam;
  isTiebreak?: LiveMatchSessionIsTiebreak;
  tiebreakPointsTeamOne?: number;
  tiebreakPointsTeamTwo?: number;
  confirmedByTeamOne?: number;
  confirmedByTeamTwo?: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  updatedAt?: string;
}